% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-paste.r
\name{as.character.paste}
\alias{as.character.paste}
\title{Extract just the paste text from a paste object}
\usage{
\method{as.character}{paste}(x, ...)
}
\arguments{
\item{x}{paste object}

\item{...}{unused}
}
\description{
Extract just the paste text from a paste object
}
