% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-recent-pastes.r
\name{get_recent_pastes}
\alias{get_recent_pastes}
\title{Get recent pastes}
\usage{
get_recent_pastes(limit = 50, lang = NULL)
}
\arguments{
\item{limit}{number of recent pastes to fetch. Limit is 500, default is 50.}

\item{lang}{limit the recent paste list to a particular language. Default is all pastes}
}
\description{
Get recent pastes
}
\references{
\href{http://pastebin.com/api_scraping_faq}{Scraping API}
}
