% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-trending-pastes.r
\name{get_trending_pastes}
\alias{get_trending_pastes}
\title{Get trending pastes}
\usage{
get_trending_pastes(pastebin_key = pastebin_api_key())
}
\arguments{
\item{pastebin_key}{pastebin API key}
}
\description{
Get trending pastes
}
\references{
\url{https://pastebin.com/api#10}
}
