% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-paste.r
\name{new_paste}
\alias{new_paste}
\title{Create a new paste}
\usage{
new_paste(x, name = NULL, format = "text", visibility = c("public",
  "unlisted", "private"), expires = "n",
  pastebin_api_key = pastebin_api_key())
}
\arguments{
\item{name}{name/title of paste}

\item{format}{hint for syntax highlighting. Defaults to \code{text}. See
\href{http://pastebin.com/api#5}{the detail page} for more info.}

\item{visibility}{one of \code{public}, \code{unlisted} or \code{private}. Defaults to \code{public}}

\item{expires}{either \code{n} for never or an abbreviated time expiration string in the form
of a digit (the "number of") and a units character \code{m} for minute(s),
\code{d} for day(s), \code{w} for week(s). Defaults to \code{n} (never). See
\href{http://pastebin.com/api#6}{the detail page} for more info.}

\item{pastebin_api_key}{pastebin API key}

\item{text}{of paste}
}
\description{
Create a new paste
}
\note{
The maximum size a paste can be is 512 kilobytes (0.5 megabytes). Pro members are
allowed to create pastes up to 10 megabytes.
}
