% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastebin-package.R
\docType{package}
\name{pastebin}
\alias{pastebin}
\alias{pastebin-package}
\title{Tools to work with the pastebin API}
\description{
Tools to work with the pastebin API
}
\author{
Bob Rudis (bob@rud.is)
}
