% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastebin-api-key.r
\name{pastebin_api_key}
\alias{pastebin_api_key}
\title{Get or set PASTEBIN_API_KEY value}
\usage{
pastebin_api_key(force = FALSE)
}
\arguments{
\item{force}{force setting a new pastebin API key for the current environment?}
}
\value{
atomic character vector containing the pastebin API key
}
\description{
The API wrapper functions in this package all rely on a pastebin API
key residing in the environment variable \code{PASTEBIN_API_KEY}. The
easiest way to accomplish this is to set it in the `\code{.Renviron}` file in your
home directory.
}
\note{
an pastebin API key is only necessary for "poster" access
}
