% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-paste.r
\name{toString.paste}
\alias{toString.paste}
\title{Extract just the paste text from a paste object}
\usage{
\method{toString}{paste}(x, ...)
}
\arguments{
\item{x}{paste object}

\item{...}{unused}
}
\description{
Extract just the paste text from a paste object
}
