% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-paste.r
\name{get_paste}
\alias{get_paste}
\title{Get raw paste data}
\usage{
get_paste(x, use_scraping_api = FALSE, include_metadata = FALSE)
}
\arguments{
\item{x}{paste id}

\item{use_scraping_api}{if a pro member, set this to \code{TRUE}, otherwise leave it \code{FALSE}
and be kind to their servers lest ye be banned.}

\item{include_metadata}{if \code{use_scraping_api} is \code{TRUE} and this is \code{TRUE}, the returned
\code{list} will include metadata}
}
\value{
a \code{list} with the paste text or the paste text plus metadata. A \code{list} is returned
to make it easier to deal with the results programmatically. Returning a \code{list}
in one call context and a \code{character} vector in another may be OK interactively
bit it creates a situation where you need to write \code{if} logic to handle
programmatically. Use \link{toString} to extract just the paste body
}
\description{
Get raw paste data
}
\references{
\href{http://pastebin.com/api_scraping_faq}{Scraping API}
}
