% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-paste-metadata.r
\name{get_paste_metadata}
\alias{get_paste_metadata}
\title{Get paste metadata}
\usage{
get_paste_metadata(paste_id)
}
\arguments{
\item{paste_id}{paste id}
}
\description{
Get paste metadata
}
\note{
This API call uses the Scraping API which requires a paid account and a white-listed IP address.
}
\references{
\href{https://pastebin.com/api_scraping_faq}{Scraping API}
}
