% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-paste-metadata.r
\name{get_paste_metadata}
\alias{get_paste_metadata}
\title{Get paste metadata}
\usage{
get_paste_metadata(x)
}
\arguments{
\item{x}{paste id}

\item{use_scraping_api}{if a pro member, set this to \code{TRUE}, otherwise leave it \code{FALSE}
and be kind to their servers lest ye be banned.}

\item{include_metadata}{if \code{use_scraping_api} is \code{TRUE} and this is \code{TRUE}, the returned
\code{list} will include metadata}
}
\description{
Get paste metadata
}
\references{
\href{http://pastebin.com/api_scraping_faq}{Scraping API}
}
