% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-paste.r
\name{new_paste}
\alias{new_paste}
\title{Create a new paste}
\usage{
new_paste(text, name = NULL, format = "text", impersonate = FALSE,
  visibility = c("public", "unlisted", "private"), expires = "n",
  pastebin_key = pastebin_api_key())
}
\arguments{
\item{text}{of paste}

\item{name}{name/title of paste}

\item{format}{hint for syntax highlighting. Defaults to \code{text}. See
\href{https://pastebin.com/api#5}{the detail page} for more info.}

\item{impersonate}{if \code{TRUE} then \code{PASTEBIN_USER} and \code{PASTEBIN_PASSWORD} \emph{must} be set
in order to generate a user key to be applied with the API key. Don't blame me,
blame \href{https://pastebin.com/api#8}{pastebin}.}

\item{visibility}{one of \code{public}, \code{unlisted} or \code{private}. Defaults to \code{public}}

\item{expires}{either \code{n} for never or an abbreviated time expiration string in the form
of a digit (the "number of") and a units character \code{m} for minute(s),
\code{d} for day(s), \code{w} for week(s). Defaults to \code{n} (never). See
\href{https://pastebin.com/api#6}{the detail page} for more info.}

\item{pastebin_key}{pastebin API key}
}
\description{
Create a new paste
}
\note{
The maximum size a paste can be is 512 kilobytes (0.5 megabytes). Pro members are
allowed to create pastes up to 10 megabytes.
}
